scripts/requirements-bg-removal.txt
# 去背景服务依赖包
rembg>=2.0.67
Pillow>=10.1.0
numpy>=1.24.3
onnxruntime>=1.16.3
requests>=2.31.0
